# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/ordens-servicos/{id}/execucoes": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Inicia a execução de uma ordem de serviço após aprovação do orçamento",
        "operationId": "executar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutarOrdemServicoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Execução da OS iniciada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ordem de serviço não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Orçamento não aprovado ou OS em estado incorreto",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos/{id}/execucoes",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
    "/orcamentos": {
      "post": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Criar novo orçamento",
        "description": "Cria um novo orçamento com peças e serviços",
        "operationId": "criarOrcamento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarOrcamentoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Orçamento criado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/orcamentos",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/orcamentos/{id}/recusados": {
      "post": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Recusar orçamento",
        "description": "Marca o orçamento como recusado",
        "operationId": "recusarOrcamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do orçamento",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Orçamento recusado com sucesso"
          },
          "404": {
            "description": "Orçamento não encontrado"
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/orcamentos/{id}/recusados",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/orcamentos/{id}/aceitos": {
      "post": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Aceitar orçamento",
        "description": "Marca o orçamento como aceito",
        "operationId": "aceitarOrcamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do orçamento",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Orçamento aceito com sucesso"
          },
          "404": {
            "description": "Orçamento não encontrado"
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/orcamentos/{id}/aceitos",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
    "/orcamentos/{id}": {
      "get": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Buscar orçamento por ID",
        "description": "Retorna os detalhes do orçamento",
        "operationId": "buscarOrcamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do orçamento",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orçamento encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Orçamento não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/orcamentos/{id}",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/orcamentos/{id}/envios": {
      "get": {
        "tags": [
          "Orçamentos"
        ],
        "summary": "Enviar orçamento em PDF",
        "description": "Gera e envia o orçamento em formato PDF para download",
        "operationId": "enviarOrcamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do orçamento",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF do orçamento gerado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Orçamento não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/orcamentos/{id}/envios",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutarOrdemServicoRequest": {
        "type": "object",
        "properties": {
          "orcamentoId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CriarOrcamentoRequest": {
        "type": "object",
        "description": "Dados para criação do orçamento",
        "properties": {
          "cpfCnpjCliente": {
            "type": "string"
          },
          "veiculoId": {
            "type": "string",
            "format": "uuid"
          },
          "pecas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PecaOrcamentoRequest"
            }
          },
          "servicosIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "cpfCnpjCliente",
          "veiculoId"
        ]
      },
      "PecaOrcamentoRequest": {
        "type": "object",
        "properties": {
          "pecaId": {
            "type": "string",
            "format": "uuid"
          },
          "quantidade": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          }
        },
        "required": [
          "pecaId"
        ]
      },
...
    }
  }
}